refactor: 폴더 이동(filter 를 common으로 이동)
기존 filter폴더를 common으로 이동
