+++ src/main/java/racingcar/common/exception/DuplicateCarNameException.java
@@ -1,0 +1,10 @@
+package racingcar.common.exception;
+
+import racingcar.common.constant.MessageConstants;
+
+public class DuplicateCarNameException extends IllegalArgumentException {
+
+    public DuplicateCarNameException() {
+        super(MessageConstants.DUPLICATE_CAR_NAME_ERROR);
+    }
+}
