+++ src/main/java/racingcar/common/filter/RacingCarListParsingFilter.java
@@ -1,8 +1,10 @@
-package racingcar.filter;
+package racingcar.common.filter;
 
 import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
 import racingcar.common.exception.InvalidCarNameException;
+import racingcar.common.exception.DuplicateCarNameException;
 import racingcar.common.util.CarListCollector;
 import racingcar.dto.RaceRequest;
 import racingcar.dto.RaceResponse;
@@ -22,6 +24,7 @@ public class RacingCarListParsingFilter implements Filter<RaceRequest, RaceRespo
 
     private RacingCarList createCarListFromRequest(String carNamesExpression) {
         List<String> carNames = parseCarNames(carNamesExpression);
+        validateCarNames(carNames); // 유효성 검사 및 중복 체크
         return carNames.stream()
                 .map(RacingCarProxy::create)
                 .collect(CarListCollector.toCarList());
@@ -30,14 +33,18 @@ public class RacingCarListParsingFilter implements Filter<RaceRequest, RaceRespo
     private List<String> parseCarNames(String expression) {
         return Arrays.stream(expression.split(NAME_DELIMITER))
                 .map(String::trim)
-                .filter(this::isValidCarName)
                 .toList();
     }
 
-    private boolean isValidCarName(String name) {
-        if (name.length() > MAX_CAR_NAME_LENGTH) {
-            throw new InvalidCarNameException();
+    private void validateCarNames(List<String> carNames) {
+        HashSet<String> uniqueNames = new HashSet<>();
+        for (String name : carNames) {
+            if (name.length() > MAX_CAR_NAME_LENGTH) {
+                throw new InvalidCarNameException();
+            }
+            if (!uniqueNames.add(name)) {
+                throw new DuplicateCarNameException(); // 중복 예외 발생
+            }
         }
-        return true;
     }
 }
