+++ src/main/java/racingcar/service/RaceServiceImpl.java
@@ -4,7 +4,7 @@ import java.util.List;
 import racingcar.common.util.StringUtil;
 import racingcar.dto.RaceRequest;
 import racingcar.dto.RaceResponse;
-import racingcar.filter.RaceFilterChain;
+import racingcar.common.filter.RaceFilterChain;
 
 public class RaceServiceImpl implements RaceService {
     private final RaceFilterChain raceFilterChain;
