+++ src/test/java/racingcar/common/filter/RaceInputValidationFilterTest.java
@@ -1,4 +1,4 @@
-package racingcar.filter;
+package racingcar.common.filter;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.common.exception.InvalidCommandException;
 import racingcar.common.exception.NegativeIterationException;
+import racingcar.common.filter.RaceFilterChain;
+import racingcar.common.filter.RaceInputValidationFilter;
 import racingcar.dto.RaceRequest;
 
 class RaceInputValidationFilterTest {
